Refactor zp to a separate module
